LPS-187797 fix circular dependency error
