+++ modules/apps/accessibility/accessibility-menu-web/package.json
@@ -8,12 +8,12 @@
 		"@clayui/modal": "3.96.0",
 		"@clayui/progress-bar": "3.65.0",
 		"@liferay/cookies-banner-web": "*",
+		"@liferay/frontend-js-react-web": "*",
 		"@liferay/frontend-js-state-web": "*",
 		"frontend-js-web": "*",
 		"prop-types": "15.7.2",
 		"react": "16.12.0"
 	},
-	"main": "js/index.js",
 	"name": "@liferay/accessibility-menu-web",
 	"private": true,
 	"scripts": {
