+++ modules/apps/frontend-js/frontend-js-react-web/package.json
@@ -1,7 +1,6 @@
 {
 	"dependencies": {
 		"@clayui/icon": "3.56.0",
-		"@liferay/accessibility-menu-web": "*",
 		"@liferay/frontend-js-state-web": "*",
 		"classnames": "2.3.1",
 		"formik": "2.1.4",
