+++ modules/apps/frontend-js/frontend-js-react-web/src/main/resources/META-INF/resources/js/index.ts
@@ -25,6 +25,7 @@ export {default as usePrevious} from './hooks/usePrevious';
 export {default as useStateSafe} from './hooks/useStateSafe';
 export {default as useThunk} from './hooks/useThunk';
 export {default as useTimeout} from './hooks/useTimeout';
+export {CONSTANTS, accessibilityMenuAtom} from './accessibilityMenu';
 
 // Egregious hack because react-dnd expects `window.process` to exist:
 //
