+++ modules/apps/frontend-js/frontend-js-react-web/types/src/main/resources/META-INF/resources/js/index.d.ts
@@ -22,3 +22,4 @@ export {default as usePrevious} from './hooks/usePrevious';
 export {default as useStateSafe} from './hooks/useStateSafe';
 export {default as useThunk} from './hooks/useThunk';
 export {default as useTimeout} from './hooks/useTimeout';
+export {CONSTANTS, accessibilityMenuAtom} from './accessibilityMenu';
